seperated into five files, renamed constants
