+++ backend/populators/tickers.js
@@ -6,8 +6,7 @@ const app = express();
 app.use(express.json());
 const router = express.Router({ mergeParams: true });
 const yahooFinance = require("yahoo-finance2").default;
-const { formatDate, getBeforeDate } = require("../lib/utils");
-const { div } = require("@tensorflow/tfjs");
+const { updateAllCompanies } = require("../lib/utils");
 
 const wait = (ms) => {
   return new Promise((resolve) => setTimeout(resolve, ms)); // needed to not go over api call limit
@@ -213,91 +212,10 @@ router.post("/industry-sector-desc-fill", async (req, res) => {
 
 // batch calls for companies, used for TC 1:
 
-const BATCH_SIZE = 100;
-const MAX_BETWEEN_TIME_PRICE = 1000 * 60 * 90; // 1.5 hr update cycle, don't want to get ip banned from yfinance -> no batch call for yfinance js unfortunately
 router.post("/", async (req, res) => {
   await updateAllCompanies();
   res.json({ message: "prices updated" });
 });
-
-const updateAllCompanies = async () => {
-  const mostRecent = await prisma.company.findFirst({
-    orderBy: {
-      lastUpdate: "desc",
-    },
-  });
-  const currentTime = new Date();
-  if (
-    Math.abs(mostRecent.lastUpdate.getTime() - currentTime.getTime()) <
-    MAX_BETWEEN_TIME_PRICE
-  ) {
-    return;
-  }
-  const allCompanies = await prisma.company.findMany();
-  const onlyTickers = allCompanies.map((value) => value.ticker);
-  let currentInd = 0;
-  while (true) {
-    // to prevent yfinance crash
-    batchSplit = onlyTickers.slice(currentInd, currentInd + BATCH_SIZE);
-    if (batchSplit.length == 0) {
-      break;
-    }
-    const prices = await yahooFinance.quote(
-      batchSplit,
-      { modules: ["price"] },
-      { validateResult: false }
-    );
-    const firstDate = formatDate(getBeforeDate(""));
-    const secondDate = formatDate(new Date());
-    const options = {
-      period1: firstDate,
-      period2: secondDate,
-      events: "dividends",
-    };
-    const dividends = [];
-    for (let ticker of batchSplit) {
-      await wait(200);
-      try {
-        const rep = (await yahooFinance.chart(ticker, options)).events
-          .dividends;
-        dividends.push(rep);
-      } catch (err) {
-        dividends.push(null);
-      }
-    }
-    currentInd += BATCH_SIZE;
-    for (let i = 0; i < prices.length; i++) {
-      const company = prices[i];
-      if (dividends[i] == null) {
-        await prisma.company.update({
-          where: {
-            ticker: company.symbol,
-          },
-          data: {
-            daily_price: company.regularMarketPrice,
-            daily_price_change: company.regularMarketChangePercent,
-            lastUpdate: new Date(),
-          },
-        });
-      } else {
-        const dividendsDateList = dividends[i].map((val) => val.date);
-        const dividendsAmountList = dividends[i].map((val) => val.amount);
-        await prisma.company.update({
-          where: {
-            ticker: company.symbol,
-          },
-          data: {
-            daily_price: company.regularMarketPrice,
-            daily_price_change: company.regularMarketChangePercent,
-            lastUpdate: new Date(),
-            dividends: dividendsAmountList,
-            dividendsDates: dividendsDateList,
-          },
-        });
-      }
-    }
-  }
-};
 
 // helper functions below
