+++ frontend/src/components/WeightingTab.jsx
@@ -1,0 +1,9 @@
+const WeightingTab = () => {
+  return (
+    <button className="bg-lime-900 mx-1 text-white m-4 hover:brightness-80 hover:scale-105 w-[80%] ml-auto mr-auto">
+      Change Weightings
+    </button>
+  );
+};
+
+export default WeightingTab;
